Refactor/#62: 카테고리 컴포넌트 분리 및 아이템 효과 구현
